requirements.txt
flask
twilio
gunicorn
python-dotenv
pypdf
faiss-cpu
# STABLE VERSIONS (Fixes 'No module named chains')
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai
langchain-text-splitters